requirements.txt
#requirements.txt

pyyaml>=6.0,<7.0
